+++ ProjectSettings/EditorBuildSettings.asset
@@ -4,21 +4,5 @@
 EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
-<<<<<<< HEAD
-  m_Scenes:
-  - enabled: 1
-    path: Assets/Scenes/MapScene.unity
-    guid: dd2f227a1642b468b89b155b65229330
-  m_configObjects:
-    com.unity.adaptiveperformance.loader_settings: {fileID: 11400000, guid: b29913d7f20e8460e95176a82f1837ad,
-      type: 2}
-    com.unity.adaptiveperformance.samsung.android.provider_settings: {fileID: 11400000,
-      guid: 543b5bd9b656d408ead88436ceb078fa, type: 2}
-    com.unity.adaptiveperformance.simulator.provider_settings: {fileID: 11400000,
-      guid: bbca45832abd2462b9e8f1cf1c6169b9, type: 2}
-=======
   m_Scenes: []
-  m_configObjects:
-    com.unity.adaptiveperformance.simulator.provider_settings: {fileID: 11400000,
-      guid: a0a5a1f4635bb48a4afc17e0ec3654c8, type: 2}
->>>>>>> main
+  m_configObjects: {}
